+++ libraries/kotlinx-metadata/jvm/ChangeLog.md
@@ -1,7 +1,10 @@
 # kotlinx-metadata-jvm
 
-## 0.1.1
+## 0.2.0
 
+- ['KT-41011`](https://youtrack.jetbrains.com/issue/41011) Using KotlinClassMetadata.Class.Writer with metadata version < 1.4 will write incorrect version requirement table
+    - Breaking change: `KotlinClassMetadata.*.Writer.write` no longer accept `metadataVersion`/`bytecodeVersion`.
+      `kotlinx-metadata-jvm` 0.2.0 will only support writing metadata of version 1.4 (which is readable by Kotlin compiler/reflection of versions 1.3 and later).
 - [`KT-42429`](https://youtrack.jetbrains.com/issue/KT-42429) Wrong interpretation of Flag.Constructor.IS_PRIMARY
     - Breaking change: `Flag.Constructor.IS_PRIMARY` is deprecated, use `Flag.Constructor.IS_SECONDARY` instead
 - [`KT-37421`](https://youtrack.jetbrains.com/issue/KT-37421) Add Flag.Class.IS_FUN for functional interfaces